refactor

+++ src/main/kotlin/com/example/ticketbookingapp/webApi/screenings/dto/ScreeningSelectionResponseDto.kt
@@ -4,5 +4,5 @@ import com.example.ticketbookingapp.webApi.dto.MovieScreeningDto
 
 data class ScreeningSelectionResponseDto(
     val elements: List<MovieScreeningDto>,
-    val totalElementsNumber: Long,
+    val totalElementsNumber: Int,
 )
